fix request slots and slim registry header
