+++ include/zoo/tools/registry.hpp
@@ -1,6 +1,6 @@
 /**
  * @file registry.hpp
- * @brief Thread-safe tool registration, schema normalization, and invocation helpers.
+ * @brief Tool registration, schema normalization, and invocation helpers.
  */
 
 #pragma once
@@ -8,6 +8,7 @@
 #include "types.hpp"
 #include <concepts>
 #include <cstddef>
+#include <exception>
 #include <functional>
 #include <initializer_list>
 #include <nlohmann/json.hpp>
@@ -18,7 +19,6 @@
 #include <tuple>
 #include <type_traits>
 #include <unordered_map>
-#include <unordered_set>
 #include <utility>
 #include <vector>
 
@@ -113,33 +113,7 @@ template <typename Tuple> consteval bool tuple_types_supported() {
     return tuple_types_supported_impl<Tuple>(std::make_index_sequence<std::tuple_size_v<Tuple>>{});
 }
 
-/**
- * @brief Builds the normalized JSON Schema object for a parameter list.
- */
-inline nlohmann::json build_parameters_schema(const std::vector<ToolParameter>& parameters) {
-    nlohmann::json properties = nlohmann::json::object();
-    nlohmann::json required = nlohmann::json::array();
-
-    for (const auto& parameter : parameters) {
-        nlohmann::json property = {{"type", tool_value_type_name(parameter.type)}};
-        if (!parameter.description.empty()) {
-            property["description"] = parameter.description;
-        }
-        if (!parameter.enum_values.empty()) {
-            property["enum"] = parameter.enum_values;
-        }
-
-        properties[parameter.name] = std::move(property);
-        if (parameter.required) {
-            required.push_back(parameter.name);
-        }
-    }
-
-    return nlohmann::json{{"type", "object"},
-                          {"properties", std::move(properties)},
-                          {"required", std::move(required)},
-                          {"additionalProperties", false}};
-}
+[[nodiscard]] nlohmann::json build_parameters_schema(const std::vector<ToolParameter>& parameters);
 
 /**
  * @brief Extracts and converts one named JSON argument.
@@ -176,238 +150,19 @@ template <typename T> nlohmann::json wrap_result(T&& value) {
     return nlohmann::json{{"result", std::forward<T>(value)}};
 }
 
-[[nodiscard]] inline Expected<ToolValueType> parse_tool_value_type(std::string_view value) {
-    if (value == "integer") {
-        return ToolValueType::Integer;
-    }
-    if (value == "number") {
-        return ToolValueType::Number;
-    }
-    if (value == "string") {
-        return ToolValueType::String;
-    }
-    if (value == "boolean") {
-        return ToolValueType::Boolean;
-    }
-
-    return std::unexpected(Error{ErrorCode::InvalidToolSchema,
-                                 "Unsupported tool parameter type: " + std::string(value)});
-}
-
-[[nodiscard]] inline bool json_matches_type(const nlohmann::json& value, ToolValueType type) {
-    switch (type) {
-    case ToolValueType::Integer:
-        return value.is_number_integer();
-    case ToolValueType::Number:
-        return value.is_number();
-    case ToolValueType::String:
-        return value.is_string();
-    case ToolValueType::Boolean:
-        return value.is_boolean();
-    }
-    return false;
-}
-
-[[nodiscard]] inline Expected<void> validate_root_schema_keys(const nlohmann::json& schema,
-                                                              const std::string& tool_name) {
-    static const std::unordered_set<std::string> kAllowedKeys = {
-        "type", "properties", "required", "additionalProperties", "description"};
-
-    for (const auto& [key, _] : schema.items()) {
-        if (kAllowedKeys.contains(key)) {
-            continue;
-        }
-        return std::unexpected(
-            Error{ErrorCode::InvalidToolSchema,
-                  "Unsupported tool schema keyword '" + key + "' for tool '" + tool_name + "'"});
-    }
-
-    return {};
-}
-
-[[nodiscard]] inline Expected<void> validate_property_schema_keys(const nlohmann::json& property,
-                                                                  const std::string& tool_name,
-                                                                  const std::string& param_name) {
-    static const std::unordered_set<std::string> kAllowedKeys = {"type", "description", "enum"};
-
-    for (const auto& [key, _] : property.items()) {
-        if (kAllowedKeys.contains(key)) {
-            continue;
-        }
-        return std::unexpected(Error{ErrorCode::InvalidToolSchema,
-                                     "Unsupported schema keyword '" + key + "' for parameter '" +
-                                         param_name + "' on tool '" + tool_name + "'"});
-    }
-
-    return {};
-}
-
-[[nodiscard]] inline Expected<void> validate_enum_values(const std::vector<nlohmann::json>& values,
-                                                         ToolValueType type,
-                                                         const std::string& tool_name,
-                                                         const std::string& param_name) {
-    for (const auto& value : values) {
-        if (!json_matches_type(value, type)) {
-            return std::unexpected(Error{ErrorCode::InvalidToolSchema,
-                                         "Enum value for parameter '" + param_name + "' on tool '" +
-                                             tool_name + "' does not match declared type '" +
-                                             tool_value_type_name(type) + "'"});
-        }
-    }
-
-    return {};
-}
-
-[[nodiscard]] inline Expected<ToolMetadata>
-normalize_manual_tool_metadata(const std::string& name, const std::string& description,
-                               const nlohmann::json& schema) {
-    if (!schema.is_object()) {
-        return std::unexpected(
-            Error{ErrorCode::InvalidToolSchema, "Tool schema must be a JSON object"});
-    }
-
-    if (auto result = validate_root_schema_keys(schema, name); !result) {
-        return std::unexpected(result.error());
-    }
-
-    auto type_it = schema.find("type");
-    if (type_it == schema.end() || !type_it->is_string() || *type_it != "object") {
-        return std::unexpected(
-            Error{ErrorCode::InvalidToolSchema,
-                  "Tool schema for '" + name + "' must declare top-level type 'object'"});
-    }
-
-    auto props_it = schema.find("properties");
-    if (props_it == schema.end() || !props_it->is_object()) {
-        return std::unexpected(
-            Error{ErrorCode::InvalidToolSchema,
-                  "Tool schema for '" + name + "' must contain an object-valued 'properties'"});
-    }
-
-    auto additional_it = schema.find("additionalProperties");
-    if (additional_it != schema.end() &&
-        (!additional_it->is_boolean() || additional_it->get<bool>())) {
-        return std::unexpected(Error{ErrorCode::InvalidToolSchema,
-                                     "Tool schema for '" + name +
-                                         "' must omit 'additionalProperties' or set it to false"});
-    }
-
-    std::vector<std::string> required_names;
-    std::unordered_set<std::string> required_lookup;
-    if (auto required_it = schema.find("required"); required_it != schema.end()) {
-        if (!required_it->is_array()) {
-            return std::unexpected(
-                Error{ErrorCode::InvalidToolSchema,
-                      "Tool schema for '" + name + "' must use an array for 'required'"});
-        }
-
-        for (const auto& value : *required_it) {
-            if (!value.is_string()) {
-                return std::unexpected(
-                    Error{ErrorCode::InvalidToolSchema,
-                          "Tool schema for '" + name + "' must use string entries in 'required'"});
-            }
-
-            const std::string required_name = value.get<std::string>();
-            if (!required_lookup.insert(required_name).second) {
-                return std::unexpected(
-                    Error{ErrorCode::InvalidToolSchema,
-                          "Tool schema for '" + name +
-                              "' contains duplicate names in 'required': " + required_name});
-            }
-
-            if (!props_it->contains(required_name)) {
-                return std::unexpected(
-                    Error{ErrorCode::InvalidToolSchema,
-                          "Tool schema for '" + name +
-                              "' lists a missing property in 'required': " + required_name});
-            }
-
-            required_names.push_back(required_name);
-        }
-    }
-
-    std::unordered_map<std::string, ToolParameter> parameters_by_name;
-    for (const auto& [param_name, property] : props_it->items()) {
-        if (!property.is_object()) {
-            return std::unexpected(
-                Error{ErrorCode::InvalidToolSchema,
-                      "Property '" + param_name + "' on tool '" + name + "' must be an object"});
-        }
-
-        if (auto result = validate_property_schema_keys(property, name, param_name); !result) {
-            return std::unexpected(result.error());
-        }
-
-        auto property_type_it = property.find("type");
-        if (property_type_it == property.end() || !property_type_it->is_string()) {
-            return std::unexpected(
-                Error{ErrorCode::InvalidToolSchema, "Property '" + param_name + "' on tool '" +
-                                                        name + "' must declare a string 'type'"});
-        }
-
-        auto type = parse_tool_value_type(property_type_it->get_ref<const std::string&>());
-        if (!type) {
-            auto error = type.error();
-            error.context = "parameter=" + param_name + ", tool=" + name;
-            return std::unexpected(std::move(error));
-        }
-
-        ToolParameter parameter;
-        parameter.name = param_name;
-        parameter.type = *type;
-        parameter.required = required_lookup.contains(param_name);
-
-        if (auto description_it = property.find("description"); description_it != property.end()) {
-            if (!description_it->is_string()) {
-                return std::unexpected(Error{ErrorCode::InvalidToolSchema,
-                                             "Property '" + param_name + "' on tool '" + name +
-                                                 "' must use a string 'description'"});
-            }
-            parameter.description = description_it->get<std::string>();
-        }
-
-        if (auto enum_it = property.find("enum"); enum_it != property.end()) {
-            if (!enum_it->is_array()) {
-                return std::unexpected(Error{ErrorCode::InvalidToolSchema,
-                                             "Property '" + param_name + "' on tool '" + name +
-                                                 "' must use an array for 'enum'"});
-            }
-            parameter.enum_values = enum_it->get<std::vector<nlohmann::json>>();
-            if (auto result =
-                    validate_enum_values(parameter.enum_values, parameter.type, name, param_name);
-                !result) {
-                return std::unexpected(result.error());
-            }
-        }
-
-        parameters_by_name.emplace(param_name, std::move(parameter));
-    }
-
-    // Build canonical parameter order: required params first (in the order
-    // listed in the "required" array), then optional params (in nlohmann::json
-    // object iteration order, which is alphabetical by default).
-    std::vector<ToolParameter> parameters;
-    parameters.reserve(parameters_by_name.size());
-
-    for (const auto& required_name : required_names) {
-        auto node = parameters_by_name.extract(required_name);
-        parameters.push_back(std::move(node.mapped()));
-    }
-    for (const auto& [param_name, _] : props_it->items()) {
-        if (!required_lookup.contains(param_name)) {
-            auto node = parameters_by_name.extract(param_name);
-            parameters.push_back(std::move(node.mapped()));
-        }
-    }
-
-    ToolMetadata metadata;
-    metadata.name = name;
-    metadata.description = description;
-    metadata.parameters = std::move(parameters);
-    metadata.parameters_schema = build_parameters_schema(metadata.parameters);
-    return metadata;
-}
+[[nodiscard]] Expected<ToolValueType> parse_tool_value_type(std::string_view value);
+[[nodiscard]] bool json_matches_type(const nlohmann::json& value, ToolValueType type);
+[[nodiscard]] Expected<void> validate_root_schema_keys(const nlohmann::json& schema,
+                                                       const std::string& tool_name);
+[[nodiscard]] Expected<void> validate_property_schema_keys(const nlohmann::json& property,
+                                                           const std::string& tool_name,
+                                                           const std::string& param_name);
+[[nodiscard]] Expected<void> validate_enum_values(const std::vector<nlohmann::json>& values,
+                                                  ToolValueType type, const std::string& tool_name,
+                                                  const std::string& param_name);
+[[nodiscard]] Expected<ToolMetadata> normalize_manual_tool_metadata(const std::string& name,
+                                                                    const std::string& description,
+                                                                    const nlohmann::json& schema);
 
 /**
  * @brief Normalizes a JSON Schema object into a parameter vector.
@@ -419,14 +174,7 @@ normalize_manual_tool_metadata(const std::string& name, const std::string& descr
  * @param schema JSON Schema object with type "object".
  * @return Normalized parameter vector in canonical order (required first, then optional).
  */
-[[nodiscard]] inline Expected<std::vector<ToolParameter>>
-normalize_schema(const nlohmann::json& schema) {
-    auto result = normalize_manual_tool_metadata("schema", "", schema);
-    if (!result) {
-        return std::unexpected(result.error());
-    }
-    return std::move(result->parameters);
-}
+[[nodiscard]] Expected<std::vector<ToolParameter>> normalize_schema(const nlohmann::json& schema);
 
 template <typename Tuple, size_t... Is>
 std::vector<ToolParameter> build_typed_parameters_impl(const std::vector<std::string>& param_names,
@@ -488,16 +236,9 @@ make_tool_definition(const std::string& name, const std::string& description,
     return ToolDefinition{std::move(metadata), std::move(handler)};
 }
 
-inline Expected<ToolDefinition> make_tool_definition(const std::string& name,
-                                                     const std::string& description,
-                                                     const nlohmann::json& schema,
-                                                     ToolHandler handler) {
-    auto metadata = normalize_manual_tool_metadata(name, description, schema);
-    if (!metadata) {
-        return std::unexpected(metadata.error());
-    }
-    return ToolDefinition{std::move(*metadata), std::move(handler)};
-}
+Expected<ToolDefinition> make_tool_definition(const std::string& name,
+                                              const std::string& description,
+                                              const nlohmann::json& schema, ToolHandler handler);
 
 } // namespace detail
 
@@ -550,14 +291,7 @@ class ToolRegistry {
      * @brief Registers a tool using a prebuilt JSON Schema and handler.
      */
     Expected<void> register_tool(const std::string& name, const std::string& description,
-                                 nlohmann::json schema, ToolHandler handler) {
-        auto definition =
-            detail::make_tool_definition(name, description, schema, std::move(handler));
-        if (!definition) {
-            return std::unexpected(definition.error());
-        }
-        return register_tool(std::move(*definition));
-    }
+                                 nlohmann::json schema, ToolHandler handler);
 
     /**
      * @brief Registers a normalized tool definition.
@@ -565,18 +299,7 @@ class ToolRegistry {
      * Existing entries with the same name are replaced in place while
      * preserving registration order.
      */
-    Expected<void> register_tool(ToolDefinition definition) {
-        auto it = index_by_name_.find(definition.metadata.name);
-        if (it != index_by_name_.end()) {
-            tools_[it->second] = std::move(definition);
-            return {};
-        }
-
-        const size_t index = tools_.size();
-        index_by_name_.emplace(definition.metadata.name, index);
-        tools_.push_back(std::move(definition));
-        return {};
-    }
+    Expected<void> register_tool(ToolDefinition definition);
 
     /**
      * @brief Registers multiple tool definitions under a single lock acquisition.
@@ -587,120 +310,54 @@ class ToolRegistry {
      * @param definitions Tool definitions to register.
      * @return Void on success.
      */
-    Expected<void> register_tools(std::vector<ToolDefinition> definitions) {
-        for (auto& definition : definitions) {
-            auto it = index_by_name_.find(definition.metadata.name);
-            if (it != index_by_name_.end()) {
-                tools_[it->second] = std::move(definition);
-            } else {
-                const size_t index = tools_.size();
-                index_by_name_.emplace(definition.metadata.name, index);
-                tools_.push_back(std::move(definition));
-            }
-        }
-        return {};
-    }
+    Expected<void> register_tools(std::vector<ToolDefinition> definitions);
 
     /**
      * @brief Reports whether a tool name is currently registered.
      */
-    bool has_tool(const std::string& name) const {
-        return index_by_name_.find(name) != index_by_name_.end();
-    }
+    bool has_tool(const std::string& name) const;
 
     /**
      * @brief Invokes a registered tool with JSON arguments.
      */
-    Expected<nlohmann::json> invoke(const std::string& name, const nlohmann::json& args) const {
-        auto it = index_by_name_.find(name);
-        if (it == index_by_name_.end()) {
-            return std::unexpected(Error{ErrorCode::ToolNotFound, "Tool not found: " + name});
-        }
-        return tools_[it->second].handler(args);
-    }
+    Expected<nlohmann::json> invoke(const std::string& name, const nlohmann::json& args) const;
 
     /**
      * @brief Returns the OpenAI-style tool schema for one registered tool.
      */
-    nlohmann::json get_tool_schema(const std::string& name) const {
-        auto metadata = get_tool_metadata(name);
-        if (!metadata) {
-            return nlohmann::json{};
-        }
-        return build_schema_json(*metadata);
-    }
+    nlohmann::json get_tool_schema(const std::string& name) const;
 
     /**
      * @brief Returns the raw normalized JSON parameter schema for a registered tool.
      */
-    std::optional<nlohmann::json> get_parameters_schema(const std::string& name) const {
-        auto metadata = get_tool_metadata(name);
-        if (!metadata) {
-            return std::nullopt;
-        }
-        return metadata->parameters_schema;
-    }
+    std::optional<nlohmann::json> get_parameters_schema(const std::string& name) const;
 
     /**
      * @brief Returns normalized metadata for a registered tool.
      */
-    std::optional<ToolMetadata> get_tool_metadata(const std::string& name) const {
-        auto it = index_by_name_.find(name);
-        if (it == index_by_name_.end()) {
-            return std::nullopt;
-        }
-        return tools_[it->second].metadata;
-    }
+    std::optional<ToolMetadata> get_tool_metadata(const std::string& name) const;
 
     /**
      * @brief Returns normalized metadata for every registered tool in registration order.
      */
-    std::vector<ToolMetadata> get_all_tool_metadata() const {
-        std::vector<ToolMetadata> metadata;
-        metadata.reserve(tools_.size());
-        for (const auto& tool : tools_) {
-            metadata.push_back(tool.metadata);
-        }
-        return metadata;
-    }
+    std::vector<ToolMetadata> get_all_tool_metadata() const;
 
     /**
      * @brief Returns schemas for every registered tool in registration order.
      */
-    nlohmann::json get_all_schemas() const {
-        nlohmann::json schemas = nlohmann::json::array();
-        for (const auto& tool : tools_) {
-            schemas.push_back(build_schema_json(tool.metadata));
-        }
-        return schemas;
-    }
+    nlohmann::json get_all_schemas() const;
 
     /**
      * @brief Returns the names of every registered tool in registration order.
      */
-    std::vector<std::string> get_tool_names() const {
-        std::vector<std::string> names;
-        names.reserve(tools_.size());
-        for (const auto& tool : tools_) {
-            names.push_back(tool.metadata.name);
-        }
-        return names;
-    }
+    std::vector<std::string> get_tool_names() const;
 
     /// Returns the number of registered tools.
-    size_t size() const {
-        return tools_.size();
-    }
+    size_t size() const;
 
   private:
     /// Converts tool metadata into the schema shape consumed by prompts.
-    static nlohmann::json build_schema_json(const ToolMetadata& metadata) {
-        return nlohmann::json{{"type", "function"},
-                              {"function",
-                               {{"name", metadata.name},
-                                {"description", metadata.description},
-                                {"parameters", metadata.parameters_schema}}}};
-    }
+    static nlohmann::json build_schema_json(const ToolMetadata& metadata);
 
     std::vector<ToolDefinition> tools_;
     std::unordered_map<std::string, size_t> index_by_name_;
